refactor(protocol): split wire surface into per-interface headers
The monolithic Types.h / Marshalling.h forced a translation unit to
recompile whenever any interface's wire types changed. Split both into
six interface groups, each with a QtCore-only *Types.h and a matching
QtDBus *Marshalling.h:

  Drag, Window (window-state), Autotile, Navigation (snap-nav),
  Zone (zone/overlay), Bridge (compositor-bridge).

registerWireTypes() moves to Registration.h. ServiceConstants.h is now
shipped by the QtCore-only PhosphorProtocol::Types target, so a consumer
that needs only the service constants no longer pulls QtDBus.

All ~40 consumers across libs/, src/, kwin-effect/ and tests/ are
reclassified to the specific interface headers they use. PhosphorProtocol.h
remains as an umbrella convenience header.

Full build clean, all 184 tests pass.
