+++ kwin-effect/plasmazoneseffect/daemon_apply.cpp
@@ -7,7 +7,8 @@
 #include <PhosphorIdentity/WindowId.h>
 #include <PhosphorProtocol/ClientHelpers.h>
 #include <PhosphorProtocol/ServiceConstants.h>
-#include <PhosphorProtocol/Marshalling.h>
+#include <PhosphorProtocol/NavigationMarshalling.h>
+#include <PhosphorProtocol/WindowMarshalling.h>
 
 #include <effect/effecthandler.h>
 #include <window.h>
