+++ kwin-effect/screenchangehandler.h
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include <PhosphorProtocol/Marshalling.h>
+#include <PhosphorProtocol/WindowMarshalling.h>
 #include <QObject>
 #include <QPointer>
 #include <QRect>
