+++ kwin-effect/snapassisthandler.cpp
@@ -8,7 +8,7 @@
 
 #include <PhosphorProtocol/ServiceConstants.h>
 #include <PhosphorProtocol/ClientHelpers.h>
-#include <PhosphorProtocol/Marshalling.h>
+#include <PhosphorProtocol/ZoneMarshalling.h>
 #include <PhosphorCompositor/SnapAssistFilter.h>
 
 #include <effect/effecthandler.h>
