+++ libs/phosphor-placement/include/PhosphorPlacement/WindowTrackingService.h
@@ -9,7 +9,8 @@
 #include <PhosphorEngine/EngineTypes.h>
 #include <PhosphorEngine/IWindowTrackingService.h>
 #include <PhosphorEngine/WindowRegistry.h>
-#include <PhosphorProtocol/Types.h>
+#include <PhosphorProtocol/WindowTypes.h>
+#include <PhosphorProtocol/ZoneTypes.h>
 #include <QObject>
 #include <QString>
 #include <QStringList>
