+++ libs/phosphor-protocol/README.md
@@ -33,20 +33,39 @@ PhosphorProtocol::Types    QtCore-only value vocabulary
 
 ## Key headers
 
+The wire surface is split per interface group so a translation unit
+recompiles only when the interface it actually uses changes. Each group
+has a QtCore-only `*Types.h` and a matching QtDBus `*Marshalling.h`:
+
+| Interface group | Types header | Marshalling header |
+|-----------------|--------------|--------------------|
+| Drag        | `DragTypes.h`       | `DragMarshalling.h` |
+| Window state | `WindowTypes.h`    | `WindowMarshalling.h` |
+| Autotile    | `AutotileTypes.h`   | `AutotileMarshalling.h` |
+| Snap navigation | `NavigationTypes.h` | `NavigationMarshalling.h` |
+| Zone / overlay | `ZoneTypes.h`    | `ZoneMarshalling.h` |
+| Compositor bridge | `BridgeTypes.h` | `BridgeMarshalling.h` |
+
+Plus the cross-cutting headers:
+
 | Header | Purpose |
 |--------|---------|
-| `PhosphorProtocol/Types.h`            | Value structs + enums that cross D-Bus (drag policy, window IDs, zone rects, navigation result types) plus `Q_DECLARE_METATYPE`. **QtCore only** — no QtDBus. |
-| `PhosphorProtocol/Marshalling.h`      | `QDBusArgument` `operator<<` / `operator>>` for the `Types.h` structs, the `HasDBusStreaming` `static_assert` guard, and `registerWireTypes()`. |
-| `PhosphorProtocol/ServiceConstants.h` | Canonical service name, object path, interface names (`org.plasmazones.Autotile`, `.Screen`, `.Overlay`, …), API version, timeouts. |
+| `PhosphorProtocol/ServiceConstants.h` | Canonical service name, object path, interface names (`org.plasmazones.Autotile`, `.Screen`, `.Overlay`, …), API version, timeouts. **QtCore only** — shipped by `::Types`. |
+| `PhosphorProtocol/Registration.h`     | `registerWireTypes()` — one-shot metatype registration; call once at startup. |
 | `PhosphorProtocol/ClientHelpers.h`    | `daemonClient()` — a `PhosphorDBus::Client` value bound to the daemon — plus thin `ClientHelpers::` wrappers (`fireAndForget`, `asyncCall`, `syncCall`, `loadSettingAsync`). |
+| `PhosphorProtocol/PhosphorProtocol.h` | Umbrella convenience header that pulls in everything; prefer the per-interface headers. |
+
+A `*Types.h` header is pure value vocabulary (`Q_DECLARE_METATYPE`, no
+QtDBus); the matching `*Marshalling.h` adds the `QDBusArgument` operators
+and the `HasDBusStreaming` `static_assert` guard.
 
 ## Design notes
 
-- **Value types are transport-free.** `Types.h` defines the structs with
-  no QtDBus dependency, so a domain library can name `MoveTargetResult`
-  without linking the bus. The `QDBusArgument` marshallers are free
-  functions in `Marshalling.h`, layered on top — the struct never knows
-  it travels over D-Bus.
+- **Value types are transport-free.** A `*Types.h` header defines its
+  structs with no QtDBus dependency, so a domain library can name
+  `MoveTargetResult` without linking the bus. The `QDBusArgument`
+  marshallers are free functions in the matching `*Marshalling.h`,
+  layered on top — the struct never knows it travels over D-Bus.
 - **One source of truth for interface names.** New compositor
   integrations depend on `ServiceConstants` instead of hand-rolling the
   string.
@@ -58,10 +77,10 @@ PhosphorProtocol::Types    QtCore-only value vocabulary
   `PhosphorDBus::Client` by value, pre-bound to `org.plasmazones` on the
   session bus. The generic call mechanics live in `phosphor-dbus`; this
   library only supplies the PlasmaZones endpoint. No singleton.
-- **Build-time marshaller check.** Add a struct to `Types.h` and forget
-  its `operator<<` / `>>` in `Marshalling.h`, and the `HasDBusStreaming`
-  `static_assert` fails the build instead of crashing D-Bus dispatch at
-  runtime.
+- **Build-time marshaller check.** Add a struct to a `*Types.h` and
+  forget its `operator<<` / `>>` in the matching `*Marshalling.h`, and the
+  `HasDBusStreaming` `static_assert` fails the build instead of crashing
+  D-Bus dispatch at runtime.
 
 ## Dependencies
 
@@ -72,4 +91,4 @@ PhosphorProtocol::Types    QtCore-only value vocabulary
 ## See also
 
 - [`phosphor-screens`](../phosphor-screens/README.md) — `DBusScreenAdaptor` implements the `.Screen` interface from these constants.
-- [`phosphor-snap-engine`](../phosphor-snap-engine/README.md) — navigation result types (`MoveTargetResult`, `FocusTargetResult`, etc.) come from `Types.h` and are linked QtDBus-free via `PhosphorProtocol::Types`.
+- [`phosphor-snap-engine`](../phosphor-snap-engine/README.md) — navigation result types (`MoveTargetResult`, `FocusTargetResult`, etc.) come from `NavigationTypes.h` and are linked QtDBus-free via `PhosphorProtocol::Types`.
