+++ libs/phosphor-protocol/include/PhosphorProtocol/AutotileTypes.h
@@ -1,0 +1,87 @@
+// SPDX-FileCopyrightText: 2026 fuddlesworth
+// SPDX-License-Identifier: LGPL-2.1-or-later
+
+#pragma once
+
+#include <PhosphorProtocol/phosphorprotocoltypes_export.h>
+
+#include <QList>
+#include <QMetaType>
+#include <QRect>
+#include <QString>
+
+namespace PhosphorProtocol {
+
+/// D-Bus struct for autotile tile requests: (siiiissbb)
+struct PHOSPHORPROTOCOLTYPES_EXPORT TileRequestEntry
+{
+    QString windowId;
+    int x = 0;
+    int y = 0;
+    int width = 0;
+    int height = 0;
+    QString zoneId;
+    QString screenId;
+    bool monocle = false;
+    bool floating = false;
+
+    QRect toRect() const
+    {
+        return QRect(x, y, width, height);
+    }
+
+    /// Returns empty QString if valid, or a human-readable description of
+    /// the invariant violation. Call at every unmarshal site to detect a
+    /// garbled payload before acting on it.
+    QString validationError() const;
+};
+
+using TileRequestList = QList<TileRequestEntry>;
+
+/// D-Bus struct for algorithm metadata: (sssbbbbdibsbb)
+struct AlgorithmInfoEntry
+{
+    QString id;
+    QString name;
+    QString description;
+    bool supportsMasterCount = false;
+    bool supportsSplitRatio = false;
+    bool centerLayout = false;
+    bool producesOverlappingZones = false;
+    double defaultSplitRatio = 0.5;
+    int defaultMaxWindows = 0;
+    bool isScripted = false;
+    QString zoneNumberDisplay;
+    bool isUserScript = false;
+    bool supportsMemory = false;
+};
+
+using AlgorithmInfoList = QList<AlgorithmInfoEntry>;
+
+/// D-Bus struct for pre-tile geometry entries: (siiiiis)
+/// Replaces the JSON blob previously returned by getPreTileGeometriesJson.
+struct PreTileGeometryEntry
+{
+    QString appId;
+    int x = 0;
+    int y = 0;
+    int width = 0;
+    int height = 0;
+    QString screenId;
+
+    QRect toRect() const
+    {
+        return QRect(x, y, width, height);
+    }
+};
+
+using PreTileGeometryList = QList<PreTileGeometryEntry>;
+
+} // namespace PhosphorProtocol
+
+Q_DECLARE_METATYPE(PhosphorProtocol::TileRequestEntry)
+Q_DECLARE_METATYPE(PhosphorProtocol::TileRequestList)
+Q_DECLARE_METATYPE(PhosphorProtocol::AlgorithmInfoEntry)
+Q_DECLARE_METATYPE(PhosphorProtocol::AlgorithmInfoList)
+Q_DECLARE_METATYPE(PhosphorProtocol::PreTileGeometryEntry)
+Q_DECLARE_METATYPE(PhosphorProtocol::PreTileGeometryList)
