+++ libs/phosphor-protocol/src/types.cpp
@@ -1,7 +1,9 @@
 // SPDX-FileCopyrightText: 2026 fuddlesworth
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
-#include <PhosphorProtocol/Types.h>
+#include <PhosphorProtocol/AutotileTypes.h>
+#include <PhosphorProtocol/BridgeTypes.h>
+#include <PhosphorProtocol/DragTypes.h>
 
 #include <QDebug>
 #include <QLatin1String>
