+++ libs/phosphor-snap-engine/include/PhosphorSnapEngine/SnapEngine.h
@@ -10,7 +10,8 @@
 #include <PhosphorEngine/IWindowTrackingService.h>
 #include <PhosphorEngine/PlacementEngineBase.h>
 #include <PhosphorLayoutApi/EdgeGaps.h>
-#include <PhosphorProtocol/Types.h>
+#include <PhosphorProtocol/NavigationTypes.h>
+#include <PhosphorProtocol/WindowTypes.h>
 #include <QObject>
 #include <QPointer>
 #include <QRect>
