+++ src/daemon/main.cpp
@@ -12,8 +12,8 @@
 #include "version.h"
 #include "rendering/zoneshaderitem.h"
 #include "vulkan_support.h"
+#include <PhosphorProtocol/Registration.h>
 #include <PhosphorProtocol/ServiceConstants.h>
-#include <PhosphorProtocol/Marshalling.h>
 #include <QGuiApplication>
 #include <QCommandLineParser>
 #include <QDBusConnection>
