+++ src/dbus/windowdragadaptor.h
@@ -4,7 +4,8 @@
 #pragma once
 
 #include "plasmazones_export.h"
-#include <PhosphorProtocol/Marshalling.h>
+#include <PhosphorProtocol/DragMarshalling.h>
+#include <PhosphorProtocol/ZoneMarshalling.h>
 #include <QDBusAbstractAdaptor>
 #include <QObject>
 #include <QString>
