+++ tests/unit/autotile/test_autotile_tile_request_validation.cpp
@@ -36,7 +36,7 @@
 #include <PhosphorTileEngine/AutotileConfig.h>
 #include <PhosphorTiles/AlgorithmRegistry.h>
 #include <PhosphorTiles/TilingState.h>
-#include <PhosphorProtocol/Types.h>
+#include <PhosphorProtocol/AutotileTypes.h>
 
 using namespace PlasmaZones;
 using namespace PhosphorTileEngine;
