+++ tests/unit/dbus/test_autotile_adaptor_panel_gate.cpp
@@ -16,7 +16,7 @@
 #include <QScopeGuard>
 #include <QSignalSpy>
 
-#include <PhosphorProtocol/Marshalling.h>
+#include <PhosphorProtocol/WindowMarshalling.h>
 
 #include <PhosphorTileEngine/AutotileEngine.h>
 #include "../helpers/AutotileTestHelpers.h"
