+++ tests/unit/dbus/test_dbus_adaptor_routing.cpp
@@ -37,7 +37,9 @@
 #include <QUuid>
 #include <QVariant>
 
-#include <PhosphorProtocol/Marshalling.h>
+#include <PhosphorProtocol/NavigationMarshalling.h>
+#include <PhosphorProtocol/Registration.h>
+#include <PhosphorProtocol/WindowMarshalling.h>
 
 using namespace PhosphorProtocol;
